requirements.txt
PyYAML==6.0.1
PyQt6==6.6.1
python-dotenv==1.0.1
anthropic==0.45.0
requests==2.31.0
elevenlabs==1.51.0
google-genai
groq
